+++ inc/staffSvg/svgSymbols.ts
@@ -203,7 +203,7 @@ const symbolRules = [
 
 	pathFrameSymbol("DOT", "M c     s -  -s- - - -s-  - z"),
 
-	conditionSymbol("NULL LARGE_RECT", elem => elem.identity.type === "rect" && elem.identity.height === 2 && elem.identity.width >= 50 && elem.rx <= 0 && elem.ry <= 0),
+	conditionSymbol("NULL LARGE_RECT", elem => elem.identity.type === "rect" && (elem.identity.height === 2 || elem.identity.height === 2.25) && elem.identity.width >= 50 && elem.rx <= 0 && elem.ry <= 0),
 	conditionSymbol("NULL COPYRIGHT", elem => elem.identity.type === "text" && /www\.lilypond\.org/.test(elem.identity.text)),
 ];